Fix thread move assignment, construct thread_impl
Fix move assignment, and add test that exposed the bug.

Also fix default constructing and destrucrting a `std::thread`, along
with a test that exposed the bug (which was a random crash due to the
`std::function` construct never being called).

Finally, atomic.hpp's `history_` is updated to be zero-initialized after
MSAN reported an uninitialized memory read.